refactor PB source
