+++ core/opengate_core/opengate_core.cpp
@@ -228,7 +228,7 @@ void init_GateSourceManager(py::module &);
 
 void init_GateGenericSource(py::module &);
 
-void init_GatePBSource(py::module &);
+void init_GatePencilBeamSource(py::module &m);
 
 void init_GateVoxelsSource(py::module &);
 
@@ -396,7 +396,7 @@ PYBIND11_MODULE(opengate_core, m) {
   init_GateVSource(m);
   init_GateSourceManager(m);
   init_GateGenericSource(m);
-  init_GatePBSource(m);
+  init_GatePencilBeamSource(m);
   init_GateVoxelsSource(m);
   init_GateGANSource(m);
   init_GateSPSPosDistribution(m);
