+++ opengate/__init__.py
@@ -47,6 +47,7 @@ from .source.SourceEngine import *
 from .source.GANSourceConditionalGenerator import *
 from .source.GANSourceConditionalPairsGenerator import *
 from .source.VoxelizedSourceConditionGenerator import *
+from .source.PencilBeamSource import *
 from .helpers_physics import *
 
 # Actor specific
