+++ opengate/helpers_tests.py
@@ -267,6 +267,7 @@ def assert_images(
     t = np.fabs((s1 - s2) / s1) * 100
     b = t < sum_tolerance
     print_test(b, f"Img sums {s1} vs {s2} : {t:.2f} %  (tol {sum_tolerance:.2f} %)")
+    is_ok = is_ok and b
 
     print(f"Image1: {info1.size} {info1.spacing} {info1.origin} {ref_filename1}")
     print(f"Image2: {info2.size} {info2.spacing} {info2.origin} {filename2}")
