+++ opengate/tests/src/test044_pbs_weight.py
@@ -58,7 +58,7 @@ phantom.material = "G4_AIR"
 phantom.color = [1, 0, 1, 1]
 
 # default source for tests (from test42)
-source = sim.add_source("PB", "mysource1")
+source = sim.add_source("PencilBeam", "mysource1")
 source.mother = "waterbox1"
 source.energy.mono = 60 * MeV
 source.particle = "proton"
@@ -113,7 +113,7 @@ phantom2.material = "G4_AIR"
 phantom2.color = [1, 0, 1, 1]
 
 # default source for tests (from test42)
-source2 = sim.add_source("PB", "mysource2")
+source2 = sim.add_source("PencilBeam", "mysource2")
 source2.mother = "waterbox2"
 source2.energy.mono = 60 * MeV
 source2.particle = "proton"
@@ -172,14 +172,19 @@ print(stat)
 # ----------------------------------------------------------------------------------------------------------------
 # tests
 
-# energy deposition: we expact the edep from source two
+# energy deposition: we extract the edep from source two
 # to be double the one of source one
 
 print("\nDifference for EDEP")
 mhd_1 = "phantom_a_1.mhd"
 mhd_2 = "phantom_a_2.mhd"
 test = gate.assert_images(
-    output_path / mhd_1, output_path / mhd_2, stat, tolerance=50, ignore_value=0
+    output_path / mhd_1,
+    output_path / mhd_2,
+    stat,
+    axis="x",
+    tolerance=50,
+    ignore_value=0,
 )
 fig1 = gate.create_2D_Edep_colorMap(output_path / mhd_1, show=False)
 fig2 = gate.create_2D_Edep_colorMap(output_path / mhd_2, show=False)
